refactor(sss): Refactor into smaller modules
